requirements.txt
python-telegram-bot
psycopg2>=2.8.5
geopy>=1.21.0